+++ __tasks__/check-yarn.js
@@ -9,7 +9,7 @@ const YARN_VERSION = '1.3.2';
         const { stdout: version } = await execa('yarn', ['-v']);
 
         if (version !== YARN_VERSION) {
-            await require('./lib/log').warn(
+            require('./log').warn(
                 `You need to update yarn to ${YARN_VERSION}`,
                 `Your installed version is ${version}`,
                 `See https://yarnpkg.com/lang/en/docs/install`,
@@ -17,7 +17,7 @@ const YARN_VERSION = '1.3.2';
             process.exit(1);
         }
     } catch (e) {
-        await require('./lib/log').warn(
+        require('./log').warn(
             `You need to install yarn@${YARN_VERSION}`,
             `https://yarnpkg.com/lang/en/docs/install`,
         );
